requirements.txt
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.0
protobuf==3.17.3
PyMySQL==1.0.2
six==1.16.0
